HIVE-4525 Support timestamps earlier than 1970 and later than 2038.

+++ ql/src/java/org/apache/hadoop/hive/ql/udf/UDFToInteger.java
@@ -183,7 +183,7 @@ public class UDFToInteger extends UDF {
     if (i == null) {
       return null;
     } else {
-      intWritable.set(i.getSeconds());
+      intWritable.set((int) i.getSeconds());
       return intWritable;
     }
   }
